dashboard/requirements.txt
streamlit==1.45.*
plotly==6.*
psycopg2-binary
pandas
pydantic-settings
